MDEV-34158: Implement the GIS function ST_Geohash
The function returns the geohash corresponding to the input values.
The GIS function ST_GeoHash takes as input:

  (longitude, latitude, max_length), OR
  (point, max_length)

The longitue parameter is a numeric value in the interval [180, -180],
the latitude is a numeric value in the interval [90, -90].
In the case of point, the x coordinate is treated as the latitude and
the y coordinate is treated as the latitude. Even in the case of a
point, the same constraints apply.
The max_length parameter is the upper limit on the resulting string
size and cannot exceed 100.

Author: StefanoPetrilli <stefanop_1999@hotmail.it>
Co-authored-by: kevincheng2 <chengyf112@gmail.com>
Co-authored-by: Catalin Besleaga <catalin.besleaga@oracle.com>
Co-authored-by: Gleb Shchepa <gleb.shchepa@oracle.com>
Co-authored-by: Tatiana Azundris Nuernberg <tatjana.nuernberg@oracle.com>
Co-authored-by: Martin Hansson <martin.hansson@oracle.com>
Co-authored-by: Deepa Dixit <deepa.dixit@oracle.com>
Co-authored-by: Hans H Melby <hans.h.melby@oracle.com>
Co-authored-by: Jens Even Berg Blomsøy <jens.even.blomsoy@oracle.com>
Co-authored-by: Erlend Dahl <erlend.dahl@oracle.com>
Co-authored-by: Norvald H. Ryeng <norvald.ryeng@oracle.com>
Co-authored-by: BennyWang <benny.wang@oracle.com>
Co-authored-by: David.Zhao <david.zhao@oracle.com>
Co-authored-by: Erik Froseth <erik.froseth@oracle.com>