Secret can now be delivered via a File by prepending FILE: magic.
Use innodb_data_encryption_filekey prepend FILE: to your option. MariaDB will
then interpret the secret as Filename, which is located at the provider url.
The size of your secret is limited to 256 characters. After reading the secret
from the file it will be overwritten such that it cannot be read after server start.